Move Pair class to util package.
